% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{deploy_and_update_status}
\alias{deploy_and_update_status}
\title{Deploy docs}
\usage{
deploy_and_update_status(
  gha_result = Sys.getenv("GHA_RESULT"),
  buildlog = Sys.getenv("BUILDLOG")
)
}
\arguments{
\item{gha_result}{result of pkgdown site}

\item{buildlog}{url to build logs}
}
\description{
Push pkgdown site to docs.ropensci.org and try to set commit status.
}
