% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{deploy_site}
\alias{deploy_site}
\title{Deploy docs}
\usage{
deploy_site(
  path = "docs-website",
  deploy_org = "ropensci-docs",
  buildlog = Sys.getenv("BUILDLOG")
)
}
\arguments{
\item{path}{directory containing the website to deploy}

\item{deploy_org}{github organization to host docs}

\item{buildlog}{url to build logs}
}
\description{
Push pkgdown site from zip file to docs.ropensci.org
}
