% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{gh_app_set_commit_status}
\alias{gh_app_set_commit_status}
\title{Set Commit Status}
\usage{
gh_app_set_commit_status(repo, sha, url, pkg, state)
}
\arguments{
\item{repo}{full repo name for example "ropensci/magick"}

\item{sha}{hash of the commit to update}

\item{url}{link to the build logs}

\item{pkg}{name of the R package}

\item{state}{string with result of rendering pkgdown}
}
\description{
Sets the commit status on a commit from a GitHub App.
Requires the `GH_APP_KEY` environment variable.
}
