% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{deploy_site}
\alias{deploy_site}
\title{Deploy docs}
\usage{
deploy_site(
  deploy_org = "ropensci-docs",
  buildlog = Sys.getenv("BUILDLOG"),
  docsfile = "docs.zip"
)
}
\arguments{
\item{deploy_org}{github organization to host docs}

\item{buildlog}{url to build logs}

\item{docsfile}{name of zip file containing site do deploy}
}
\description{
Push pkgdown site from zip file to docs.ropensci.org
}
